orchestrai_google_push/requirements.txt
# Gmail Push Notification Requirements
# Required packages for the Gmail Push notification webhook system

# Web framework
flask==3.0.0

# Google Cloud and Gmail API
google-cloud-pubsub==2.18.4
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP requests
requests==2.31.0

# Production WSGI server (optional)
gunicorn==21.2.0

# Logging and monitoring (optional)
structlog==23.1.0